Conflicts:
	Server_Seetreet/src/com/seetreet/controller/ContentProviderController.java
	Server_Seetreet/src/com/seetreet/dao/MongoDAO.java
	Server_Seetreet/src/com/seetreet/util/C.java

+++ Server_Seetreet/src/com/seetreet/http/HttpControl.java
@@ -29,12 +29,11 @@ public class HttpControl {
 			for(int i = 0;i<size;i++){
 				JSONObject existObject = jArray.getJSONObject(i);
 				// true이면 없는거, false면 기존에 있는거				
-				boolean resultContent = MongoDAO.checkPublicApiContentId(existObject.getLong("contentid"));
+				boolean resultContent = MongoDAO.checkPublicApiContentId(existObject.getInt("contentid"));
 
 				if(resultContent != false){
 					// 새로 삽입
 					JSONObject obj = HttpCall.getContentObject(existObject);
-					
 					ProviderBean prov = new ProviderBean(obj);
 					BasicDBObject provObj = MongoDAO.insertPublicProvider(prov);
